requirements-vendor.txt
Django==1.9
django-webpack-loader
djangorestframework
